+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Responsive adjustments
 
+- Allow cover/contain images in collaborator galleries
+
 
 ## [0.4.0] - 2025-03-20